requirements.txt
attrs==16.3.0
mock==2.0.0
numpy==1.12.0
olefile==0.44
pandas==0.19.2
pbr==2.0.0
Pillow==4.0.0
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
wheel==0.24.0
